JobberCRM/requirements.txt
fastapi
uvicorn
PyJWT==2.10.1
requests
python-dotenv
pydantic
pydantic[email]
aiohttp
psycopg2-binary
fastapi-limiter 
setuptools>=56.0.0
redis
Sqlalchemy 
